deleted color and old db, started diesel integration
