+++ .env
@@ -1,0 +1,3 @@
+APP_URL=auth.sqlite3
+AUTH_URL=auth.sqlite3
+KEY_URL=key.sqlite3
