--- color/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "color"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-colored = "2.1.0"
