--- database/.env
@@ -1,2 +1,0 @@
-REMOTE_DATABASE_URL=remote.sqlite3 
-DATABASE_URL=db.sqlite3
