--- database/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "database"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-colored = "2.1.0"
-diesel = {version="2.1.5", features=["sqlite"]}
-dotenvy = "0.15"
-tokio = {version="1.17", features=["full"]}
