--- database/src/product.rs
@@ -1,12 +1,0 @@
-/*
- *
- * product.rs
- * 
- * Copyright (c) 2023-2024 (authors)
- *
- * All rights reserved
- *
- * TALGS is distributed under the () license, see LICENSE for details
- * 
- * */
-
