+++ diesel.toml
@@ -1,0 +1,14 @@
+# For documentation on how to configure this file,
+# see https://diesel.rs/guides/configuring-diesel-cli
+
+[print_schema]
+file = "src/schema.rs"
+custom_type_derives = ["diesel::query_builder::QueryId", "Clone"]
+
+[print_schema.auth]
+file = "src/schema/auth.rs"
+with_docs = true
+filter = { only_tables = ["users", "groups"] }
+
+[migrations_directory]
+dir = "/home/EduardoStarZ/apps/rust/TALGS/migrations"
