+++ src/schema/auth.rs
@@ -1,0 +1,64 @@
+// @generated automatically by Diesel CLI.
+
+diesel::table! {
+    /// Representation of the `groups` table.
+    ///
+    /// (Automatically generated by Diesel.)
+    groups (id) {
+        /// The `id` column of the `groups` table.
+        ///
+        /// Its SQL type is `Integer`.
+        ///
+        /// (Automatically generated by Diesel.)
+        id -> Integer,
+        /// The `name` column of the `groups` table.
+        ///
+        /// Its SQL type is `Text`.
+        ///
+        /// (Automatically generated by Diesel.)
+        name -> Text,
+    }
+}
+
+diesel::table! {
+    /// Representation of the `users` table.
+    ///
+    /// (Automatically generated by Diesel.)
+    users (id) {
+        /// The `id` column of the `users` table.
+        ///
+        /// Its SQL type is `Integer`.
+        ///
+        /// (Automatically generated by Diesel.)
+        id -> Integer,
+        /// The `name` column of the `users` table.
+        ///
+        /// Its SQL type is `Text`.
+        ///
+        /// (Automatically generated by Diesel.)
+        name -> Text,
+        /// The `password` column of the `users` table.
+        ///
+        /// Its SQL type is `Text`.
+        ///
+        /// (Automatically generated by Diesel.)
+        password -> Text,
+        /// The `group` column of the `users` table.
+        ///
+        /// Its SQL type is `Integer`.
+        ///
+        /// (Automatically generated by Diesel.)
+        group -> Integer,
+        /// The `cpf` column of the `users` table.
+        ///
+        /// Its SQL type is `Text`.
+        ///
+        /// (Automatically generated by Diesel.)
+        cpf -> Text,
+    }
+}
+
+diesel::allow_tables_to_appear_in_same_query!(
+    groups,
+    users,
+);
